+++ scripts/prepReference.sh
@@ -28,6 +28,7 @@
 #   20141019-scd: Use the configuration parameter environment variable.
 #   20141020-scd: Check for empty target files.
 #   20141030-scd: Fix Python 2.6 compatibility issue when logging RAM size.
+#   20150109-scd: Log the Grid Engine job ID.
 #Notes:
 #
 #Bugs:
@@ -59,7 +60,9 @@ usage()
 echo "# Command           : $0 $@"
 echo "# Working Directory : $(pwd)"
 if [[ "$PBS_JOBID" != "" ]]; then
-echo "# \$PBS_JOBID        : $PBS_JOBID"
+echo "# Job ID            : $PBS_JOBID"
+elif [[ "$JOB_ID" != "" ]]; then
+echo "# Job ID            : $JOB_ID"
 fi
 echo "# Hostname          :" $(hostname)
 echo "# RAM               :" $(python -c 'from __future__ import print_function; import psutil; import locale; locale.setlocale(locale.LC_ALL, ""); print("%s MB" % locale.format("%d", psutil.virtual_memory().total / 1024 / 1024, grouping=True))')